Write test to cover all cases for CommentToBookService

+++ src/main/java/com/macalsandair/library/comment/CommentToBookService.java
@@ -48,7 +48,13 @@ public class CommentToBookService {
     }
 
     public void deleteComment(Long id) {
-        commentToBookRepository.deleteById(id);
+    	if (commentToBookRepository.findById(id).isPresent()) {
+            commentToBookRepository.deleteById(id);
+    	}
+    	else {
+    		throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, "Comment not found: " + id);
+    	}
     }
 
     public List<CommentToBook> findByAuthor(Long authorId) {
